.utils/requirement.txt
opencc~=1.1.9
python-frontmatter~=1.1.0
pip~=23.2.1
wheel~=0.41.2
setuptools~=68.2.0
